Enable `bf16` support in the benchmark scripts (#5293)
* Enable bf16 support for benchmark

* Enable bf16 support for kernel, points and inference

* Support autocast both for cpu and cuda

* Update benchmark/citation/train_eval.py

* Update benchmark/points/train_eval.py

* Update benchmark/kernel/main_performance.py

* Update CHANGELOG.md

+++ torch_geometric/nn/conv/eg_conv.py
@@ -131,7 +131,7 @@ class EGConv(MessagePassing):
                     edge_index, symnorm_weight = gcn_norm(  # yapf: disable
                         edge_index, None, num_nodes=x.size(self.node_dim),
                         improved=False, add_self_loops=self.add_self_loops,
-                        flow=self.flow)
+                        flow=self.flow, dtype=x.dtype)
                     if self.cached:
                         self._cached_edge_index = (edge_index, symnorm_weight)
                 else:
@@ -143,7 +143,7 @@ class EGConv(MessagePassing):
                     edge_index = gcn_norm(  # yapf: disable
                         edge_index, None, num_nodes=x.size(self.node_dim),
                         improved=False, add_self_loops=self.add_self_loops,
-                        flow=self.flow)
+                        flow=self.flow, dtype=x.dtype)
                     if self.cached:
                         self._cached_adj_t = edge_index
                 else:
